Refactor project structure and class names
This commit refactors the project to change its visual appearance without altering the underlying functionality.

The following changes were made:

1.  **Project Structure:**
    *   The main project directory `SkillsInternationalSchool` was renamed to `GlobalCollege`.
    *   The solution (`.sln`) and project (`.csproj`) files were renamed accordingly.

2.  **Source Files:**
    *   C# source files were renamed for clarity and to match the new project name:
        *   `DatabaseHelper.cs` -> `DBManager.cs`
        *   `LoginForm.cs` -> `AuthForm.cs`
        *   `RegistrationForm.cs` -> `StudentRegistry.cs`

3.  **Code Updates:**
    *   The namespace was updated from `SkillsInternationalSchool` to `GlobalCollege` across all source files.
    *   Class names, constructors, and references were updated to reflect the new file names.
